requirements.txt
attrs==19.3.0
cffi==1.14.0
click==7.1.2
cryptography==2.9.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.1
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==2.20.1
more-itertools==8.3.0
packaging==20.4
pluggy==0.13.1
py==1.8.1
pycparser==2.20
PyMySQL==0.9.3
pyparsing==2.4.7
pytest==5.4.2
python-dotenv==0.13.0
six==1.14.0
SQLAlchemy==1.3.17
wcwidth==0.1.9
Werkzeug==1.0.1
zipp==3.1.0
